Add tests for GifPlayerPresenter
I would like to create a wrapper over NotificationCenter to work with closures,
but it will be time consuming to implement one right now and I don't wonna copy
it from somewhere.
Another way to call selector that is saved as addObserverArgs is to use:
```
NSThread(target: presenter!,
	selector:mockNotifier.addObserverArgs!.selector,
	object: nil).main()
```
but I don't feel like this is a good way.