Add raw ApplicationLayer business logic
* The accplication layer mainly consists of two components:
- GifSearchFlow: This class controlls the flow. Which screens, in which order,
and in which way are presented
- GifRepository: Is a repository that will provide the necessary data from the
GifAPI, it can be either cached one, or remote one

* Besides these two classes a OfflineFakeGifRepository was created. This can be
used to easily test the flow and application withouth making requests to the
GiphyAPI(consuming them, since there is a limited number of requests that can
be made).

* Normaly I would cover all this code with tests, since it's mostly business
logic. But since it does not going in production right now, I will leave
it for later.